+++ src/Support/Facades/GraphQL.php
@@ -5,6 +5,8 @@ namespace Rebing\GraphQL\Support\Facades;
 
 use GraphQL\Error\Error;
 use GraphQL\Executor\ExecutionResult;
+use GraphQL\Type\Definition\NonNull;
+use GraphQL\Type\Definition\NullableType;
 use GraphQL\Type\Definition\ObjectType;
 use GraphQL\Type\Definition\Type;
 use GraphQL\Type\Schema;
@@ -16,7 +18,7 @@ use Rebing\GraphQL\Support\OperationParams;
  * @method static array execute(string $schemaName, OperationParams $operationParams, $rootValue = null, $contextValue = null)
  * @method static array query(string $query, ?array $params = null, array $opts = [])
  * @method static ExecutionResult queryAndReturnResult(string $query, ?array $params = null, array $opts = [])
- * @method static Type type(string $name, bool $fresh = false)
+ * @method static (NullableType&Type)|NonNull type(string $name, bool $fresh = false)
  * @method static Type paginate(string $typeName, string $customName = null)
  * @method static Type simplePaginate(string $typeName, string $customName = null)
  * @method static array<string,object|string> getTypes()